requirements.txt
aiounittest==1.4.2
appdirs==1.4.4
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
msgpack==1.0.8
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyee==11.1.0
PyJWT==2.8.0
pyparsing==3.1.2
pyppeteer==2.0.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
websockets==10.4
wrapt==1.16.0
zipp==3.19.2
